requirements.txt
adal==1.2.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
azure-common==1.1.16
azure-eventhub==1.2.0
azure-iothub-device-client==1.4.4
azure-iothub-service-client==1.4.4
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Bower==1.3.0
Flask-GoogleCharts==0.0.3
Flask-GoogleMaps==0.2.6
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
msrest==0.6.2
msrestazure==0.5.1
oauthlib==2.1.0
pluggy==0.8.0
py==1.7.0
pycparser==2.19
PyJWT==1.6.4
pymongo==3.7.2
pytest==4.0.1
python-dateutil==2.7.5
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
SQLAlchemy==1.2.14
uamqp==1.1.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
